refactor: moved ipfs and ens utils and validations to common folder
