+++ src/components/ActionsBuilder/SupportedActions/UpdateENSContent/UpdateENSContentSummary.tsx
@@ -8,7 +8,7 @@ import {
 import Summary from '../common/Summary';
 import { useUpdateEnsContent } from 'hooks/Guilds/useUpdateEnsContent';
 import { useEnsName, useNetwork } from 'wagmi';
-import { convertToIpfsHash, getIpfsUrl, isSupportedChainId } from './utils';
+import { convertToIpfsHash, getIpfsUrl, isSupportedChainId } from 'utils/ipfs';
 import { useTranslation } from 'react-i18next';
 import { ActionViewProps } from '..';
 import { useENSContentHash } from 'hooks/Guilds/ens/useENSPublicResolverContract';
